+++ packages/blueprints-integration/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.49.5-0](https://github.com/nrkno/tv-automation-server-core/compare/v1.49.4...v1.49.5-0) (2024-03-08)
+
+**Note:** Version bump only for package @sofie-automation/blueprints-integration
+
+
+
+
+
 ## [1.49.4](https://github.com/nrkno/tv-automation-server-core/compare/v1.49.3...v1.49.4) (2024-01-03)
 
 **Note:** Version bump only for package @sofie-automation/blueprints-integration